rename db and logger config files
