+++ src/middleware/auth.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from "express";
 import jwt from "jsonwebtoken";
-import { JWT_SECRET } from "../configs/db.config";
+import { JWT_SECRET } from "../configs/dbConfig";
 
 interface AuthRequest extends Request {
   user?: string | object;
